text_extractor/requirements.txt

fastapi
uvicorn[standard]
pypdf
python-docx
sentence-transformers
faiss-cpu
httpx
motor
python-dotenv
